Examples/Seeding: move algorithm to track finding module
