+++ Examples/Io/Performance/CMakeLists.txt
@@ -11,7 +11,7 @@ target_include_directories(
 target_link_libraries(
   ActsExamplesIoPerformance
   PUBLIC ActsExamplesFramework
-  PRIVATE ActsCore ActsExamplesSeeding  ActsPluginDigitization ROOT::Core ROOT::Tree)
+  PRIVATE ActsCore ROOT::Core ROOT::Tree)
 
 install(
   TARGETS ActsExamplesIoPerformance
